Refactor, update tests.
