+++ lib/did-io.js
@@ -3,66 +3,6 @@
  */
 import {VERIFICATION_RELATIONSHIPS} from './constants.js';
 
-export class DidResolver {
-  constructor(options = {}) {
-    this.methods = options.methods || {};
-  }
-
-  use(methodName, driver) {
-    this.methods[methodName] = driver;
-  }
-
-  async get({did, ...options} = {}) {
-    const method = this._methodForDid(did);
-    return method.get({did, ...options});
-  }
-
-  /**
-   * @param {object} options - Options hashmap.
-   * @param {object} options.didDocument - DID Document to update.
-   * @param {object} [options.updateOptions] - Options to pass to the update
-   *   operations.
-   *
-   * @returns {Promise<object>} - Returns results of the update operation.
-   */
-  async update({didDocument = {}, ...updateOptions} = {}) {
-    const method = this._methodForDid(didDocument.did);
-    return method.update(updateOptions);
-  }
-
-  /**
-   * Registers a DID Document, method-specific.
-   *
-   * TODO: Not sure if this is top level / universal?
-   *
-   * @param {object} options - Options hashmap.
-   * @param {object} options.didDocument - DID Document to register.
-   * @param {object} [options.registerOptions] - Options to pass to the register
-   *   operations.
-   *
-   * @returns {Promise<object>} - Returns results of the registration operation.
-   */
-  async register({didDocument = {}, ...registerOptions} = {}) {
-    const method = this._methodForDid(didDocument.did);
-    return method.register(registerOptions);
-  }
-
-  /**
-   * @param {string} did - DID uri.
-   *
-   * @returns {object} - DID Method driver.
-   * @private
-   */
-  _methodForDid(did) {
-    const {prefix} = _parseDid(did);
-    const method = this.methods[prefix];
-    if(!method) {
-      throw new Error(`Driver for DID ${did} not found.`);
-    }
-    return method;
-  }
-}
-
 /**
  * Tests whether this DID Document contains a verification relationship
  * between the subject and a method id, for a given purpose.
@@ -257,14 +197,14 @@ export function _methodById({doc, methodId}) {
  * Parses the DID into various component (currently, only cares about prefix).
  *
  * @example
- * _parseDid('did:v1:test:nym');
+ * parseDid({did: 'did:v1:test:nym'});
  * // -> {prefix: 'v1'}
  *
  * @param {string} did - DID uri.
  *
  * @returns {{prefix: string}} Returns the method prefix (without `did:`).
  */
-export function _parseDid(did) {
+export function parseDid({did}) {
   if(!did) {
     throw new TypeError('DID cannot be empty.');
   }
