+++ lib/example-did-driver.js
