+++ lib/main.js
@@ -1,0 +1,13 @@
+/*!
+ * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
+ */
+export {
+  approvesMethodFor,
+  findVerificationMethod,
+  initKeys,
+  parseDid
+} from './did-io.js';
+
+export {MultiResolver} from './MultiResolver.js';
+
+export {VERIFICATION_RELATIONSHIPS} from './constants.js';
