--- src/main.js
@@ -1,7 +1,0 @@
-/*!
- * Copyright (c) 2020-2021 Digital Bazaar, Inc. All rights reserved.
- */
-export {
-  DidResolver
-} from './DidResolver.js';
-export {VERIFICATION_RELATIONSHIPS} from './constants.js';
