+++ test/did-io.spec.js
@@ -5,11 +5,11 @@ import chai from 'chai';
 chai.should();
 const {expect} = chai;
 
-import {DidResolver} from '../';
+// import {MultiResolver} from '../';
 // Exported only for testing
 import {
-  findVerificationMethod, approvesMethodFor, _parseDid
-} from '../src/DidResolver.js';
+  findVerificationMethod, approvesMethodFor, parseDid
+} from '../';
 
 const MOCK_KEY = {
   id: 'did:ex:123#abcd',
@@ -18,29 +18,29 @@ const MOCK_KEY = {
   publicKeyMultibase: '...'
 };
 
-describe('_parseDid', () => {
+describe('parseDid', () => {
   it('should return main did method identifier', async () => {
-    const {prefix} = _parseDid('did:v1:test:nym:abcd');
+    const {prefix} = parseDid({did: 'did:v1:test:nym:abcd'});
     expect(prefix).to.equal('v1');
   });
 });
 
-describe('didIo resolver instance', () => {
-  const didIo = new DidResolver();
-
-  describe('get()', () => {
-    it('should error if no DID is passed', async () => {
-      let error;
-      try {
-        await didIo.get();
-      } catch(e) {
-        error = e;
-      }
-
-      expect(error.message).to.equal('DID cannot be empty.');
-    });
-  });
-});
+// describe('didIo resolver instance', () => {
+//   const didIo = new MultiResolver();
+//
+//   describe('get()', () => {
+//     it('should error if no DID is passed', async () => {
+//       let error;
+//       try {
+//         await didIo.get();
+//       } catch(e) {
+//         error = e;
+//       }
+//
+//       expect(error.message).to.equal('DID cannot be empty.');
+//     });
+//   });
+// });
 
 describe('didIo utility functions', () => {
   describe('findVerificationMethod', () => {
